adding new service entities and moving the interfaces for cert authorities in UI
